+++ products/jbrowse-react-circular-genome-view/stories/GettingStarted.mdx
@@ -6,11 +6,10 @@ import { Volvox, ShowTrack } from './JBrowseCircularGenomeView.stories'
 # Welcome!
 
 This site contains the [Storybook](https://storybook.js.org/) docs for the
-[JBrowse 2 React Circular Genome View
-Component](https://www.npmjs.com/package/@jbrowse/react-circular-genome-view).
+[JBrowse 2 React Circular Genome View Component](https://www.npmjs.com/package/@jbrowse/react-circular-genome-view).
 The component makes it possible to embed an interactive genome browser in any
-React application. The Storybook docs make it possible to provide an
-interactive demonstration of how the component looks and works.
+React application. The Storybook docs make it possible to provide an interactive
+demonstration of how the component looks and works.
 
 ## Using the docs
 
@@ -19,9 +18,9 @@ This is an example of what a story looks like:
 <Story of={Volvox} />
 
 You can test out the various features of the component, and open tracks and
-interact with it as you would in your app. You can also click on the code
-button on the bottom right to see what the code looks like for the component
-being viewed.
+interact with it as you would in your app. You can also click on the code button
+on the bottom right to see what the code looks like for the component being
+viewed.
 
 ## An overview of the CGV API
 
@@ -29,9 +28,9 @@ The CGV component relies on the same
 [configuration](https://jbrowse.org/jb2/docs/config_guide) as the main JB2 app,
 with a few noteworthy differences.
 
-First, the component can be configured using a standalone JSON file like JB2,
-or can be configured with JavaScript objects in the source code of the app
-where it is being used, as shown
+First, the component can be configured using a standalone JSON file like JB2, or
+can be configured with JavaScript objects in the source code of the app where it
+is being used, as shown
 [here](https://github.com/GMOD/jbrowse-components/blob/HEAD/products/jbrowse-react-circular-genome-view/docs/example.md).
 
 Second, while this is the structure of a JB2 configuration file:
@@ -78,7 +77,5 @@ for what is possible with the component.
 
 Here are a few more resources that may help:
 
-- [Code
-example](https://github.com/GMOD/jbrowse-components/blob/HEAD/products/jbrowse-react-circular-genome-view/docs/example.md)
-- [Story source
-code](https://github.com/GMOD/jbrowse-components/tree/main/products/jbrowse-react-circular-genome-view/stories)
+- [Code example](https://github.com/GMOD/jbrowse-components/blob/HEAD/products/jbrowse-react-circular-genome-view/docs/example.md)
+- [Story source code](https://github.com/GMOD/jbrowse-components/tree/main/products/jbrowse-react-circular-genome-view/stories)
