+++ products/jbrowse-react-circular-genome-view/stories/examples/index.ts
@@ -1,0 +1,3 @@
+export * from './Volvox'
+export * from './ShowTrack'
+export * from './Human'
