+++ products/jbrowse-react-linear-genome-view/stories/DefaultSessions.mdx
@@ -1,13 +1,13 @@
-import { Meta, Story } from '@storybook/blocks';
+import { Meta, Story } from '@storybook/blocks'
 import { DefaultSession } from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Default Sessions" />
 
 # Default sessions
 
-When using JBrowse 2, information about the current app state such as what
-views are open is stored in a session. It is possible to configure a JB2 app
-with a `defaultSession`, which will be loaded when the app first is loaded.
+When using JBrowse 2, information about the current app state such as what views
+are open is stored in a session. It is possible to configure a JB2 app with a
+`defaultSession`, which will be loaded when the app first is loaded.
 
 Here is an example LGV with a long reads alignment track loaded in the
 `defaultSession`:
@@ -84,9 +84,9 @@ sessions is the feature for exporting sessions in the main JBrowse 2 app (File
 the app. This feature can also be used to graphically generate the desired view
 configuration for the React LGV component.
 
-To create a default session using JB2, load whatever tracks and data you want
-to be displayed into a session with a single linear genome view open. Export
-the session from JB2 and save it as `session.json` in your app.
+To create a default session using JB2, load whatever tracks and data you want to
+be displayed into a session with a single linear genome view open. Export the
+session from JB2 and save it as `session.json` in your app.
 
 To use the view settings in your app, you can create the `defaultSession` like
 this:
