+++ products/jbrowse-react-linear-genome-view/stories/GettingStarted.mdx
@@ -1,9 +1,9 @@
-import { Meta, Story } from '@storybook/blocks';
+import { Meta, Story } from '@storybook/blocks'
 import {
   OneLinearGenomeView,
   UsingLocObject,
   DisableAddTrack,
-  WithShowTrack
+  WithShowTrack,
 } from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Getting Started" />
