+++ products/jbrowse-react-linear-genome-view/stories/NextJsUsage.mdx
@@ -1,4 +1,4 @@
-import { Meta } from '@storybook/blocks';
+import { Meta } from '@storybook/blocks'
 
 <Meta title="Next.js Usage" />
