+++ products/jbrowse-react-linear-genome-view/stories/NextstrainDemo.mdx
@@ -1,4 +1,4 @@
-import { Meta, Story } from '@storybook/blocks';
+import { Meta, Story } from '@storybook/blocks'
 import { NextstrainExample } from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Nextstrain SARS-CoV-2 demo" />
@@ -179,7 +179,7 @@ large-ish JSON file because of the usage of `FromConfigAdapter`
             "start": 26252,
             "end": 26253,
             "uniqueId": "26252"
-          },
+          }
           /*more scores for the entropy track here */
         ]
       }
@@ -237,5 +237,4 @@ large-ish JSON file because of the usage of `FromConfigAdapter`
     }
   }
 }
-
 ```
