+++ products/jbrowse-react-linear-genome-view/stories/TextSearching.mdx
@@ -1,5 +1,8 @@
-import { Meta, Story } from '@storybook/blocks';
-import { WithAggregateTextSearching, WithPerTrackTextSearching } from './JBrowseLinearGenomeView.stories'
+import { Meta, Story } from '@storybook/blocks'
+import {
+  WithAggregateTextSearching,
+  WithPerTrackTextSearching,
+} from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Text Searching" />
