+++ products/jbrowse-react-linear-genome-view/stories/Theming.mdx
@@ -1,4 +1,4 @@
-import { Meta, Story } from '@storybook/blocks';
+import { Meta, Story } from '@storybook/blocks'
 import { WithCustomTheme } from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Theming" />
