+++ products/jbrowse-react-linear-genome-view/stories/UsingPlugins.mdx
@@ -1,5 +1,8 @@
-import { Meta, Story } from '@storybook/blocks';
-import { WithInlinePlugins, WithExternalPlugin } from './JBrowseLinearGenomeView.stories'
+import { Meta, Story } from '@storybook/blocks'
+import {
+  WithInlinePlugins,
+  WithExternalPlugin,
+} from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Using Plugins" />
