+++ products/jbrowse-react-linear-genome-view/stories/UsingWebWorker.mdx
@@ -1,4 +1,4 @@
-import { Meta, Story } from '@storybook/blocks';
+import { Meta, Story } from '@storybook/blocks'
 import { WithWebWorker } from './JBrowseLinearGenomeView.stories'
 
 <Meta title="Using WebWorker RPC" />
